captionning/backend/requirements.txt
fastapi
uvicorn
python-multipart
transformers
accelerate
torch
torchvision
bitsandbytes
google-genai
Pillow
python-dotenv
